requirements.txt
Flask==3.1.1
requests==2.32.4
python-dotenv==1.1.1
sib-api-v3-sdk
mux-python
gunicorn

# These are sub-dependencies and optional:
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
urllib3==2.5.0
Werkzeug==3.1.3
